+++ main-service/src/main/java/ru/practicum/ewm/event/dto/EventShortDto.java
@@ -25,5 +25,5 @@ public class EventShortDto {
     CategoryDto category;
     Boolean paid;
     Long views;
-    Integer confirmedRequests;
+    Long confirmedRequests;
 }
